# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pomodoro Application Api Server.",
        "title": "Pomodoro API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/jobs/accessed-days": {
            "get": {
                "description": "Get total days user has accessed on pomodoro from date to date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get Days Accessed",
                "parameters": [
                    {
                        "description": "Get days",
                        "name": "DateToDate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.GetStatisticRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get days accessed",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal serever error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/jobs/focused-minutes": {
            "get": {
                "description": "Get total minutes user has spent on pomodoro from date to date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get Minutes Focused",
                "parameters": [
                    {
                        "description": "Get minutes",
                        "name": "DateToDate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.GetStatisticRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get minutes focused",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal serever error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/jobs/pomodoros": {
            "get": {
                "description": "Get pomodoros from date to date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "List Pomodoros",
                "parameters": [
                    {
                        "description": "Get pomodoros",
                        "name": "GetPomos",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.GetPomodorosRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Pomodoro"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal serever error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new pomodoro",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Create New Pomodoro",
                "parameters": [
                    {
                        "description": "New pomodoro",
                        "name": "NewPomo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.CreatePomodoroRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Pomodoro"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not found user_id or type_id",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal serever error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/jobs/update-user-setting": {
            "put": {
                "description": "Update user setting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Update User Setting",
                "parameters": [
                    {
                        "description": "Update user setting",
                        "name": "userSetting",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.UpdateUserSettingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update user setting successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal serever error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.Pomodoro": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "focus_degree": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "task_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "type_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "delivery.CreatePomodoroRequest": {
            "type": "object",
            "required": [
                "focus_degree",
                "type_id"
            ],
            "properties": {
                "focus_degree": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                },
                "task_id": {
                    "type": "integer"
                },
                "type_id": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "delivery.GetPomodorosRequest": {
            "type": "object",
            "required": [
                "from_date",
                "limit",
                "page",
                "to_date"
            ],
            "properties": {
                "from_date": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer",
                    "minimum": 1
                },
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "to_date": {
                    "type": "string"
                }
            }
        },
        "delivery.GetStatisticRequest": {
            "type": "object",
            "required": [
                "from_date",
                "to_date"
            ],
            "properties": {
                "from_date": {
                    "type": "string"
                },
                "to_date": {
                    "type": "string"
                }
            }
        },
...
        "delivery.UpdateUserSettingRequest": {
            "type": "object",
            "required": [
                "alarm_sound",
                "repeat_alarm",
                "username"
            ],
            "properties": {
                "alarm_sound": {
                    "type": "string"
                },
                "repeat_alarm": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 1
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}